requirements.txt
streamlit==1.46.0
langchain==0.3.0
langchain-openai==0.2.2
openai==1.47.0
python-dotenv==1.1.0